Backend for editListing
Attempt to complete de backend for editListing
