--- web-app/src/pages/CreateListingPage.tsx
@@ -1,15 +1,0 @@
-
-import EditListingForm from '../components/EditListingForm'
-
-const CreateListingPage = () => {
-  return (
-    <div>
-    
-    <EditListingForm/>
-      
-        
-    </div>
-  )
-}
-
-export default CreateListingPage
